[issue-244-refactor-network-handling-code] new NetworkMessage.cpp
 - Moved all the inlined functions of the space into NetworkMessage.cpp
 - Added NetworkMessage.cpp to Makefile.am
 - Missing functions popped into main.cpp and were fixed by including the
   Network:: prefix on each call. I think it's a lot cooler now. Thanks @zander.

 I still want to see what else can be brought from main.cpp into NetworkMessage .h/.cpp that doesn't need to be there.

Still getting one build error.
```
src/script/bitcoinconsensus.cpp:88:114: error: no matching constructor for initialization of 'TransactionSignatureChecker'
        return VerifyScript(tx.vin[nIn].scriptSig, CScript(scriptPubKey, scriptPubKey + scriptPubKeyLen), flags, TransactionSignatureChecker(&tx, nIn), NULL);
```
